refactor tests from tpath_test.go to tree_test.go and struct_test.go
